rt-app: temporary fix for uclamp syscall limitatioins
Currently it's not possible to change clamp values by not specifying a valid
policy and priority via the sched_setparams() syscall available since Linux
v5.3.

Let's add this temporary fix by reading the reusing the current task policy
and priority when we ask for uclamp changes.

This patch can be reverted once the above syscall wil be fixed.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>