barrier: fix waiting event with muli instances tasks
The events of  thetask object are parsed only once even if several
instances will be then created.
barrier.waiting must add all instances of the task that will sync.
Similar init can't be done for fork event has we don't know in advance
how many forkees will be created which means that fork and barrier can't
be used in the same task description. The solution would be to set an init
value to the barrier resource in the "resources" object but that's a
dedicated dev.

Signed-off-by: Vincent Guittot <vincent.guittot@linaro.org>